krakbot/research/requirements.txt
pandas>=2.2,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.5,<2.0
matplotlib>=3.8,<4.0
pyarrow>=16.0,<20.0
SQLAlchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
PyYAML>=6.0,<7.0
joblib>=1.4,<2.0
